+++ enterprise_extensions/gp_kernels.py
@@ -1,0 +1,192 @@
+# -*- coding: utf-8 -*-
+from __future__ import (absolute_import, division,
+                        print_function, unicode_literals)
+import numpy as np
+import scipy.stats
+from collections import OrderedDict
+
+import enterprise
+from enterprise.signals import parameter
+from enterprise.signals import selections
+from enterprise.signals import signal_base
+import enterprise.signals.signal_base as base
+from enterprise.signals import white_signals
+from enterprise.signals import gp_signals
+from enterprise.signals import deterministic_signals
+from enterprise.signals import utils
+from enterprise import constants as const
+
+from enterprise_extensions import model_utils
+import enterprise_extensions.enterprise_base as eb
+
+
+# linear interpolation basis in time with nu^-2 scaling
+@signal_base.function
+def linear_interp_basis_dm(toas, freqs, dt=30*86400):
+
+    # get linear interpolation basis in time
+    U, avetoas = utils.linear_interp_basis(toas, dt=dt)
+
+    # scale with radio frequency
+    Dm = (1400/freqs)**2
+
+    return U * Dm[:, None], avetoas
+
+# linear interpolation basis in time with nu^-4 scaling
+@signal_base.function
+def linear_interp_basis_scattering(toas, freqs, dt=30*86400):
+
+    # get linear interpolation basis in time
+    U, avetoas = utils.linear_interp_basis(toas, dt=dt)
+
+    # scale with radio frequency
+    Dm = (1400/freqs)**4
+
+    return U * Dm[:, None], avetoas
+
+# linear interpolation in radio frequcny
+@signal_base.function
+def linear_interp_basis_freq(freqs, df=64):
+
+    return utils.linear_interp_basis(freqs, dt=df)
+
+# DMX-like signal with Gaussian prior
+@signal_base.function
+def dmx_ridge_prior(avetoas, log10_sigma=-7):
+    sigma = 10**log10_sigma
+    return sigma**2 * np.ones_like(avetoas)
+
+# quasi-periodic kernel for DM
+@signal_base.function
+def periodic_kernel(avetoas, log10_sigma=-7, log10_ell=2, log10_gam_p=0, log10_p=0):
+
+    r = np.abs(avetoas[None, :] - avetoas[:, None])
+
+    # convert units to seconds
+    sigma = 10**log10_sigma
+    l = 10**log10_ell * 86400
+    p = 10**log10_p * 3.16e7
+    gam_p = 10**log10_gam_p
+    d = np.eye(r.shape[0]) * (sigma/500)**2
+    K = sigma**2 * np.exp(-r**2/2/l**2 - gam_p*np.sin(np.pi*r/p)**2) + d
+    return K
+
+# squared-exponential kernel for FD
+@signal_base.function
+def se_kernel(avefreqs, log10_sigma=-7, log10_lam=3):
+
+    tm = np.abs(avefreqs[None, :] - avefreqs[:, None])
+
+    lam = 10**log10_lam
+    sigma = 10**log10_sigma
+    d = np.eye(tm.shape[0]) * (sigma/500)**2
+    return sigma**2 * np.exp(-tm**2/2/lam) + d
+
+# squared-exponential kernel for DM
+@signal_base.function
+def se_dm_kernel(avetoas, log10_sigma=-7, log10_ell=2):
+
+    r = np.abs(avetoas[None, :] - avetoas[:, None])
+
+    # Convert everything into seconds
+    l = 10**log10_ell * 86400
+    sigma = 10**log10_sigma
+    d = np.eye(r.shape[0]) * (sigma/500)**2
+    K = sigma**2 * np.exp(-r**2/2/l**2) + d
+    return K
+
+# quantization matrix in time and radio frequency to cut down on the kernel size.
+@signal_base.function
+def get_tf_quantization_matrix(toas, freqs, dt=30*86400, df=None, dm=False, dm_idx=2):
+    if df is None:
+        dfs = [(600, 1000), (1000, 1900), (1900, 3000), (3000, 5000)]
+    else:
+        fmin = freqs.min()
+        fmax = freqs.max()
+        fs = np.arange(fmin, fmax+df, df)
+        dfs = [(fs[ii], fs[ii+1]) for ii in range(len(fs)-1)]
+
+    Us, avetoas, avefreqs, masks = [], [], [], []
+    for rng in dfs:
+        mask = np.logical_and(freqs>=rng[0], freqs<rng[1])
+        if any(mask):
+            masks.append(mask)
+            U, _ = utils.create_quantization_matrix(toas[mask],
+                                                    dt=dt, nmin=1)
+            avetoa = np.array([toas[mask][idx.astype(bool)].mean()
+                               for idx in U.T])
+            avefreq = np.array([freqs[mask][idx.astype(bool)].mean()
+                                for idx in U.T])
+            Us.append(U)
+            avetoas.append(avetoa)
+            avefreqs.append(avefreq)
+
+    nc = np.sum(U.shape[1] for U in Us)
+    U = np.zeros((len(toas), nc))
+    avetoas = np.concatenate(avetoas)
+    idx = np.argsort(avetoas)
+    avefreqs = np.concatenate(avefreqs)
+    nctot = 0
+    for ct, mask in enumerate(masks):
+        Umat = Us[ct]
+        nn = Umat.shape[1]
+        U[mask, nctot:nn+nctot] = Umat
+        nctot += nn
+
+    if dm:
+         weights = (1400/freqs)**dm_idx
+    else:
+        weights = np.ones_like(freqs)
+
+    return U[:, idx] * weights[:, None], {'avetoas': avetoas[idx],
+                                          'avefreqs': avefreqs[idx]}
+
+# kernel is the product of a quasi-periodic time kernel and
+# a rational-quadratic frequency kernel.
+@signal_base.function
+def tf_kernel(labels, log10_sigma=-7, log10_ell=2, log10_gam_p=0,
+              log10_p=0, log10_ell2=4, log10_alpha_wgt=0):
+
+    avetoas = labels['avetoas']
+    avefreqs = labels['avefreqs']
+
+    r = np.abs(avetoas[None, :] - avetoas[:, None])
+    r2 = np.abs(avefreqs[None, :] - avefreqs[:, None])
+
+    # convert units to seconds
+    sigma = 10**log10_sigma
+    l = 10**log10_ell * 86400
+    l2 = 10**log10_ell2
+    p = 10**log10_p * 3.16e7
+    gam_p = 10**log10_gam_p
+    alpha_wgt = 10**log10_alpha_wgt
+
+    d = np.eye(r.shape[0]) * (sigma/500)**2
+    Kt = sigma**2 * np.exp(-r**2/2/l**2 - gam_p*np.sin(np.pi*r/p)**2)
+    Kv = (1+r2**2/2/alpha_wgt/l2**2)**(-alpha_wgt)
+
+    return Kt * Kv + d
+
+# kernel is the product of a squared-exponential time kernel and
+# a rational-quadratic frequency kernel
+@signal_base.function
+def sf_kernel(labels, log10_sigma=-7, log10_ell=2,
+              log10_ell2=4, log10_alpha_wgt=0):
+
+    avetoas = labels['avetoas']
+    avefreqs = labels['avefreqs']
+
+    r = np.abs(avetoas[None, :] - avetoas[:, None])
+    r2 = np.abs(avefreqs[None, :] - avefreqs[:, None])
+
+    # Convert everything into seconds
+    l = 10**log10_ell * 86400
+    sigma = 10**log10_sigma
+    l2 = 10**log10_ell2
+    alpha_wgt = 10**log10_alpha_wgt
+
+    d = np.eye(r.shape[0]) * (sigma/500)**2
+    Kt = sigma**2 * np.exp(-r**2/2/l**2)
+    Kv = (1+r2**2/2/alpha_wgt/l2**2)**(-alpha_wgt)
+
+    return Kt * Kv + d
