+++ enterprise_extensions/models.py
@@ -1,3 +1,4 @@
+# -*- coding: utf-8 -*-
 from __future__ import (absolute_import, division,
                         print_function, unicode_literals)
 import numpy as np
@@ -17,1085 +18,6 @@ from enterprise import constants as const
 
 from enterprise_extensions import model_utils
 import enterprise_extensions.enterprise_base as eb
-
-#### Extra model components not part of base enterprise ####
-
-# timing model delay
-@signal_base.function
-def tm_delay(residuals, t2pulsar, tmparams_orig, tmparams, which='all'):
-    """
-    Compute difference in residuals due to perturbed timing model.
-
-    :param residuals: original pulsar residuals from Pulsar object
-    :param t2pulsar: libstempo pulsar object
-    :param tmparams_orig: dictionary of TM parameter tuples, (val, err)
-    :param tmparams: new timing model parameters, rescaled to be in sigmas
-    :param which: option to have all or only named TM parameters varied
-
-    :return: difference between new and old residuals in seconds
-    """
-
-    if which == 'all': keys = tmparams_orig.keys()
-    else: keys = which
-
-    # grab original timing model parameters and errors in dictionary
-    orig_params = np.array([tmparams_orig[key] for key in keys])
-
-    # put varying parameters into dictionary
-    tmparams_rescaled = np.atleast_1d(np.double(orig_params[:,0] +
-                                                tmparams * orig_params[:,1]))
-    tmparams_vary = OrderedDict(zip(keys, tmparams_rescaled))
-
-    # set to new values
-    t2pulsar.vals(tmparams_vary)
-    new_res = np.double(t2pulsar.residuals().copy())
-
-    # remmeber to set values back to originals
-    t2pulsar.vals(OrderedDict(zip(keys,
-                                  np.atleast_1d(np.double(orig_params[:,0])))))
-
-    # Return the time-series for the pulsar
-    return new_res - residuals
-
-# linear interpolation basis in time with nu^-2 scaling
-@signal_base.function
-def linear_interp_basis_dm(toas, freqs, dt=30*86400):
-
-    # get linear interpolation basis in time
-    U, avetoas = utils.linear_interp_basis(toas, dt=dt)
-
-    # scale with radio frequency
-    Dm = (1400/freqs)**2
-
-    return U * Dm[:, None], avetoas
-
-# linear interpolation basis in time with nu^-4 scaling
-@signal_base.function
-def linear_interp_basis_scattering(toas, freqs, dt=30*86400):
-
-    # get linear interpolation basis in time
-    U, avetoas = utils.linear_interp_basis(toas, dt=dt)
-
-    # scale with radio frequency
-    Dm = (1400/freqs)**4
-
-    return U * Dm[:, None], avetoas
-
-# linear interpolation in radio frequcny
-@signal_base.function
-def linear_interp_basis_freq(freqs, df=64):
-
-    return utils.linear_interp_basis(freqs, dt=df)
-
-# DMX-like signal with Gaussian prior
-@signal_base.function
-def dmx_ridge_prior(avetoas, log10_sigma=-7):
-    sigma = 10**log10_sigma
-    return sigma**2 * np.ones_like(avetoas)
-
-# quasi-periodic kernel for DM
-@signal_base.function
-def periodic_kernel(avetoas, log10_sigma=-7, log10_ell=2, log10_gam_p=0, log10_p=0):
-
-    r = np.abs(avetoas[None, :] - avetoas[:, None])
-
-    # convert units to seconds
-    sigma = 10**log10_sigma
-    l = 10**log10_ell * 86400
-    p = 10**log10_p * 3.16e7
-    gam_p = 10**log10_gam_p
-    d = np.eye(r.shape[0]) * (sigma/500)**2
-    K = sigma**2 * np.exp(-r**2/2/l**2 - gam_p*np.sin(np.pi*r/p)**2) + d
-    return K
-
-# squared-exponential kernel for FD
-@signal_base.function
-def se_kernel(avefreqs, log10_sigma=-7, log10_lam=3):
-
-    tm = np.abs(avefreqs[None, :] - avefreqs[:, None])
-
-    lam = 10**log10_lam
-    sigma = 10**log10_sigma
-    d = np.eye(tm.shape[0]) * (sigma/500)**2
-    return sigma**2 * np.exp(-tm**2/2/lam) + d
-
-# squared-exponential kernel for DM
-@signal_base.function
-def se_dm_kernel(avetoas, log10_sigma=-7, log10_ell=2):
-
-    r = np.abs(avetoas[None, :] - avetoas[:, None])
-
-    # Convert everything into seconds
-    l = 10**log10_ell * 86400
-    sigma = 10**log10_sigma
-    d = np.eye(r.shape[0]) * (sigma/500)**2
-    K = sigma**2 * np.exp(-r**2/2/l**2) + d
-    return K
-
-# quantization matrix in time and radio frequency to cut down on the kernel size.
-@signal_base.function
-def get_tf_quantization_matrix(toas, freqs, dt=30*86400, df=None, dm=False, dm_idx=2):
-    if df is None:
-        dfs = [(600, 1000), (1000, 1900), (1900, 3000), (3000, 5000)]
-    else:
-        fmin = freqs.min()
-        fmax = freqs.max()
-        fs = np.arange(fmin, fmax+df, df)
-        dfs = [(fs[ii], fs[ii+1]) for ii in range(len(fs)-1)]
-
-    Us, avetoas, avefreqs, masks = [], [], [], []
-    for rng in dfs:
-        mask = np.logical_and(freqs>=rng[0], freqs<rng[1])
-        if any(mask):
-            masks.append(mask)
-            U, _ = utils.create_quantization_matrix(toas[mask],
-                                                    dt=dt, nmin=1)
-            avetoa = np.array([toas[mask][idx.astype(bool)].mean()
-                               for idx in U.T])
-            avefreq = np.array([freqs[mask][idx.astype(bool)].mean()
-                                for idx in U.T])
-            Us.append(U)
-            avetoas.append(avetoa)
-            avefreqs.append(avefreq)
-
-    nc = np.sum(U.shape[1] for U in Us)
-    U = np.zeros((len(toas), nc))
-    avetoas = np.concatenate(avetoas)
-    idx = np.argsort(avetoas)
-    avefreqs = np.concatenate(avefreqs)
-    nctot = 0
-    for ct, mask in enumerate(masks):
-        Umat = Us[ct]
-        nn = Umat.shape[1]
-        U[mask, nctot:nn+nctot] = Umat
-        nctot += nn
-
-    if dm:
-         weights = (1400/freqs)**dm_idx
-    else:
-        weights = np.ones_like(freqs)
-
-    return U[:, idx] * weights[:, None], {'avetoas': avetoas[idx],
-                                          'avefreqs': avefreqs[idx]}
-
-# kernel is the product of a quasi-periodic time kernel and
-# a rational-quadratic frequency kernel.
-@signal_base.function
-def tf_kernel(labels, log10_sigma=-7, log10_ell=2, log10_gam_p=0,
-              log10_p=0, log10_ell2=4, log10_alpha_wgt=0):
-
-    avetoas = labels['avetoas']
-    avefreqs = labels['avefreqs']
-
-    r = np.abs(avetoas[None, :] - avetoas[:, None])
-    r2 = np.abs(avefreqs[None, :] - avefreqs[:, None])
-
-    # convert units to seconds
-    sigma = 10**log10_sigma
-    l = 10**log10_ell * 86400
-    l2 = 10**log10_ell2
-    p = 10**log10_p * 3.16e7
-    gam_p = 10**log10_gam_p
-    alpha_wgt = 10**log10_alpha_wgt
-
-    d = np.eye(r.shape[0]) * (sigma/500)**2
-    Kt = sigma**2 * np.exp(-r**2/2/l**2 - gam_p*np.sin(np.pi*r/p)**2)
-    Kv = (1+r2**2/2/alpha_wgt/l2**2)**(-alpha_wgt)
-
-    return Kt * Kv + d
-
-# kernel is the product of a squared-exponential time kernel and
-# a rational-quadratic frequency kernel
-@signal_base.function
-def sf_kernel(labels, log10_sigma=-7, log10_ell=2,
-              log10_ell2=4, log10_alpha_wgt=0):
-
-    avetoas = labels['avetoas']
-    avefreqs = labels['avefreqs']
-
-    r = np.abs(avetoas[None, :] - avetoas[:, None])
-    r2 = np.abs(avefreqs[None, :] - avefreqs[:, None])
-
-    # Convert everything into seconds
-    l = 10**log10_ell * 86400
-    sigma = 10**log10_sigma
-    l2 = 10**log10_ell2
-    alpha_wgt = 10**log10_alpha_wgt
-
-    d = np.eye(r.shape[0]) * (sigma/500)**2
-    Kt = sigma**2 * np.exp(-r**2/2/l**2)
-    Kv = (1+r2**2/2/alpha_wgt/l2**2)**(-alpha_wgt)
-
-    return Kt * Kv + d
-
-@signal_base.function
-def chrom_exp_decay(toas, freqs, log10_Amp=-7, sign_param=-1.0,
-                    t0=54000, log10_tau=1.7, idx=2):
-    """
-    Chromatic exponential-dip delay term in TOAs.
-
-    :param t0: time of exponential minimum [MJD]
-    :param tau: 1/e time of exponential [s]
-    :param log10_Amp: amplitude of dip
-    :param sign_param: sign of waveform
-    :param idx: index of chromatic dependence
-
-    :return wf: delay time-series [s]
-    """
-    t0 *= const.day
-    tau = 10**log10_tau * const.day
-    ind = np.where(toas > t0)[0]
-    wf = 10**log10_Amp * np.heaviside(toas - t0, 1)
-    wf[ind] *= np.exp(- (toas[ind] - t0) / tau)
-
-    return np.sign(sign_param) * wf * (1400 / freqs) ** idx
-
-@signal_base.function
-def chrom_exp_cusp(toas, freqs, log10_Amp=-7, sign_param=-1.0,
-                   t0=54000, log10_tau_pre=1.7, log10_tau_post=1.7,
-                   symmetric=False, idx=2):
-    """
-    Chromatic exponential-cusp delay term in TOAs.
-
-    :param t0: time of exponential minimum [MJD]
-    :param tau_pre: 1/e time of exponential before peak [s]
-    :param tau_post: 1/e time of exponential after peak[s]
-    :param symmetric: whether or not tau_pre = tau_post
-    :param log10_Amp: amplitude of cusp
-    :param sign_param: sign of waveform
-    :param idx: index of chromatic dependence
-
-    :return wf: delay time-series [s]
-    """
-    t0 *= const.day
-    if symmetric:
-        tau = 10**log10_tau_pre * const.day
-        ind_pre = np.where(toas < t0)[0]
-        ind_post = np.where(toas > t0)[0]
-        wf_pre = 10**log10_Amp * (1 - np.heaviside(toas - t0, 1))
-        wf_pre[ind_pre] *= np.exp(- (t0 - toas[ind_pre]) / tau)
-        wf_post = 10**log10_Amp * np.heaviside(toas - t0, 1)
-        wf_post[ind_post] *= np.exp(- (toas[ind_post] - t0) / tau)
-        wf = wf_pre + wf_post
-
-    else:
-        tau_pre = 10**log10_tau_pre * const.day
-        tau_post = 10**log10_tau_post * const.day
-        ind_pre = np.where(toas < t0)[0]
-        ind_post = np.where(toas > t0)[0]
-        wf_pre = 10**log10_Amp * (1 - np.heaviside(toas - t0, 1))
-        wf_pre[ind_pre] *= np.exp(- (t0 - toas[ind_pre]) / tau_pre)
-        wf_post = 10**log10_Amp * np.heaviside(toas - t0, 1)
-        wf_post[ind_post] *= np.exp(- (toas[ind_post] - t0) / tau_post)
-        wf = wf_pre + wf_post
-
-    return np.sign(sign_param) * wf * (1400 / freqs) ** idx
-
-@signal_base.function
-def chrom_dual_exp_cusp(toas, freqs, t0=54000, sign_param=-1.0,
-                        log10_Amp_1=-7, log10_tau_pre_1=1.7, log10_tau_post_1=1.7,
-                        log10_Amp_2=-7, log10_tau_pre_2=1.7, log10_tau_post_2=1.7,
-                        symmetric=False, idx1=2, idx2=4):
-    """
-    Chromatic exponential-cusp delay term in TOAs.
-
-    :param t0: time of exponential minimum [MJD]
-    :param tau_pre: 1/e time of exponential before peak [s]
-    :param tau_post: 1/e time of exponential after peak[s]
-    :param symmetric: whether or not tau_pre = tau_post
-    :param log10_Amp: amplitude of cusp
-    :param sign_param: sign of waveform
-    :param idx: index of chromatic dependence
-
-    :return wf: delay time-series [s]
-    """
-    t0 *= const.day
-    ind_pre = np.where(toas < t0)[0]
-    ind_post = np.where(toas > t0)[0]
-    if symmetric:
-        tau_1 = 10**log10_tau_pre_1 * const.day
-        wf_1_pre = 10**log10_Amp_1 * (1 - np.heaviside(toas - t0, 1))
-        wf_1_pre[ind_pre] *= np.exp(- (t0 - toas[ind_pre]) / tau_1)
-        wf_1_post = 10**log10_Amp_1 * np.heaviside(toas - t0, 1)
-        wf_1_post[ind_post] *= np.exp(- (toas[ind_post] - t0) / tau_1)
-        wf_1 = wf_1_pre + wf_1_post
-
-        tau_2 = 10**log10_tau_pre_2 * const.day
-        wf_2_pre = 10**log10_Amp_2 * (1 - np.heaviside(toas - t0, 1))
-        wf_2_pre[ind_pre] *= np.exp(- (t0 - toas[ind_pre]) / tau_2)
-        wf_2_post = 10**log10_Amp_2 * np.heaviside(toas - t0, 1)
-        wf_2_post[ind_post] *= np.exp(- (toas[ind_post] - t0) / tau_2)
-        wf_2 = wf_2_pre + wf_2_post
-
-    else:
-        tau_1_pre = 10**log10_tau_pre_1 * const.day
-        tau_1_post = 10**log10_tau_post_1 * const.day
-        wf_1_pre = 10**log10_Amp_1 * (1 - np.heaviside(toas - t0, 1))
-        wf_1_pre[ind_pre] *= np.exp(- (t0 - toas[ind_pre]) / tau_1_pre)
-        wf_1_post = 10**log10_Amp_1 * np.heaviside(toas - t0, 1)
-        wf_1_post[ind_post] *= np.exp(- (toas[ind_post] - t0) / tau_1_post)
-        wf_1 = wf_1_pre + wf_1_post
-
-        tau_2_pre = 10**log10_tau_pre_2 * const.day
-        tau_2_post = 10**log10_tau_post_2 * const.day
-        wf_2_pre = 10**log10_Amp_2 * (1 - np.heaviside(toas - t0, 1))
-        wf_2_pre[ind_pre] *= np.exp(- (t0 - toas[ind_pre]) / tau_2_pre)
-        wf_2_post = 10**log10_Amp_2 * np.heaviside(toas - t0, 1)
-        wf_2_post[ind_post] *= np.exp(- (toas[ind_post] - t0) / tau_2_post)
-        wf_2 = wf_2_pre + wf_2_post
-
-    return np.sign(sign_param) * ( wf_1 * (1400 / freqs) ** idx1 + wf_2 * (1400 / freqs) ** idx2)
-
-@signal_base.function
-def chrom_yearly_sinusoid(toas, freqs, log10_Amp=-7, phase=0, idx=2):
-    """
-    Chromatic annual sinusoid.
-
-    :param log10_Amp: amplitude of sinusoid
-    :param phase: initial phase of sinusoid
-    :param idx: index of chromatic dependence
-
-    :return wf: delay time-series [s]
-    """
-
-    wf = 10**log10_Amp * np.sin( 2 * np.pi * const.fyr * toas + phase)
-    return wf * (1400 / freqs) ** idx
-
-@signal_base.function
-def chromatic_quad_basis(toas, freqs, idx=4):
-    """
-    Basis for chromatic quadratic function.
-
-    :param idx: index of chromatic dependence
-
-    :return ret: normalized quadratic basis matrix [Ntoa, 3]
-    """
-    ret = np.zeros((len(toas), 3))
-    t0 = (toas.max() + toas.min()) / 2
-    for ii in range(3):
-        ret[:, ii] = (toas-t0) ** (ii) * (1400/freqs) ** idx
-    norm = np.sqrt(np.sum(ret**2, axis=0))
-    return ret/norm, np.ones(3)
-
-@signal_base.function
-def chromatic_quad_prior(toas):
-    """
-    Prior for chromatic quadratic function.
-
-    :return prior: prior-range for quadratic coefficients
-    """
-    return np.ones(3) * 1e80
-
-@signal_base.function
-def dmx_delay(toas, freqs, dmx_ids, **kwargs):
-    """
-    Delay in DMX model of DM variations.
-
-    :param dmx_ids: dictionary of DMX data for each pulsar from parfile
-    :param kwargs: dictionary of enterprise DMX parameters
-
-    :return wf: DMX signal
-    """
-    wf = np.zeros(len(toas))
-    dmx = kwargs
-    for dmx_id in dmx_ids:
-        mask = np.logical_and(toas >= (dmx_ids[dmx_id]['DMX_R1'] - 0.01) * 86400.,
-                              toas <= (dmx_ids[dmx_id]['DMX_R2'] + 0.01) * 86400.)
-        wf[mask] += dmx[dmx_id] / freqs[mask]**2 / const.DM_K / 1e12
-    return wf
-
-#### Model component building blocks ####
-
-def timing_block(tmparam_list=['RAJ', 'DECJ', 'F0', 'F1',
-                               'PMRA', 'PMDEC', 'PX']):
-    """
-    Returns the timing model block of the model
-    :param tmparam_list: a list of parameters to vary in the model
-    """
-    # default 5-sigma prior above and below the parfile mean
-    tm_params = parameter.Uniform(-5.0, 5.0, size=len(tmparam_list))
-
-    # timing model
-    tm_func = tm_delay(tmparams=tm_params, which=tmparam_list)
-    tm = deterministic_signals.Deterministic(tm_func, name='timing model')
-
-    return tm
-
-def white_noise_block(vary=False, inc_ecorr=False, gp_ecorr=False,
-                      efac1=False, select='backend', name=None):
-    """
-    Returns the white noise block of the model:
-
-        1. EFAC per backend/receiver system
-        2. EQUAD per backend/receiver system
-        3. ECORR per backend/receiver system
-
-    :param vary:
-        If set to true we vary these parameters
-        with uniform priors. Otherwise they are set to constants
-        with values to be set later.
-    :param inc_ecorr:
-        include ECORR, needed for NANOGrav channelized TOAs
-    :param gp_ecorr:
-        whether to use the Gaussian process model for ECORR
-    :param efac1:
-        use a strong prior on EFAC = Normal(mu=1, stdev=0.1)
-    """
-
-    if select == 'backend':
-        # define selection by observing backend
-        backend = selections.Selection(selections.by_backend)
-        # define selection by nanograv backends
-        backend_ng = selections.Selection(selections.nanograv_backends)
-    else:
-        # define no selection
-        backend = selections.Selection(selections.no_selection)
-
-
-    # white noise parameters
-    if vary:
-        if efac1:
-            efac = parameter.Normal(1.0, 0.1)
-        else:
-            efac = parameter.Uniform(0.01, 10.0)
-        equad = parameter.Uniform(-8.5, -5)
-        if inc_ecorr:
-            ecorr = parameter.Uniform(-8.5, -5)
-    else:
-        efac = parameter.Constant()
-        equad = parameter.Constant()
-        if inc_ecorr:
-            ecorr = parameter.Constant()
-
-    # white noise signals
-    ef = white_signals.MeasurementNoise(efac=efac,
-                                        selection=backend, name=name)
-    eq = white_signals.EquadNoise(log10_equad=equad,
-                                  selection=backend, name=name)
-    if inc_ecorr:
-        if gp_ecorr:
-            ec = gp_signals.EcorrBasisModel(log10_ecorr=ecorr,
-                                            selection=backend_ng, name=name)
-        else:
-            ec = white_signals.EcorrKernelNoise(log10_ecorr=ecorr,
-                                                selection=backend_ng, name=name)
-
-    # combine signals
-    if inc_ecorr:
-        s = ef + eq + ec
-    elif not inc_ecorr:
-        s = ef + eq
-
-    return s
-
-def red_noise_block(psd='powerlaw', prior='log-uniform', Tspan=None,
-                    components=30, gamma_val=None, coefficients=False,
-                    select=None, modes=None, wgts=None,
-                    break_flat=False, break_flat_fq=None):
-    """
-    Returns red noise model:
-        1. Red noise modeled as a power-law with 30 sampling frequencies
-    :param psd:
-        PSD function [e.g. powerlaw (default), turnover, spectrum, tprocess]
-    :param prior:
-        Prior on log10_A. Default if "log-uniform". Use "uniform" for
-        upper limits.
-    :param Tspan:
-        Sets frequency sampling f_i = i / Tspan. Default will
-        use overall time span for indivicual pulsar.
-    :param components:
-        Number of frequencies in sampling of red noise
-    :param gamma_val:
-        If given, this is the fixed slope of the power-law for
-        powerlaw, turnover, or tprocess red noise
-    :param coefficients: include latent coefficients in GP model?
-    """
-    # red noise parameters that are common
-    if psd in ['powerlaw', 'powerlaw_genmodes', 'turnover',
-               'tprocess', 'tprocess_adapt']:
-        # parameters shared by PSD functions
-        if prior == 'uniform':
-            log10_A = parameter.LinearExp(-20, -11)
-        elif prior == 'log-uniform' and gamma_val is not None:
-            if np.abs(gamma_val - 4.33) < 0.1:
-                log10_A = parameter.Uniform(-20, -11)
-            else:
-                log10_A = parameter.Uniform(-20, -11)
-        else:
-            log10_A = parameter.Uniform(-20, -11)
-
-        if gamma_val is not None:
-            gamma = parameter.Constant(gamma_val)
-        else:
-            gamma = parameter.Uniform(0, 7)
-
-        # different PSD function parameters
-        if psd == 'powerlaw':
-            pl = utils.powerlaw(log10_A=log10_A, gamma=gamma)
-        elif psd == 'powerlaw_genmodes':
-            pl = powerlaw_genmodes(log10_A=log10_A, gamma=gamma, wgts=wgts)
-        elif psd == 'turnover':
-            kappa = parameter.Uniform(0, 7)
-            lf0 = parameter.Uniform(-9, -7)
-            pl = utils.turnover(log10_A=log10_A, gamma=gamma,
-                                 lf0=lf0, kappa=kappa)
-        elif psd == 'tprocess':
-            df = 2
-            alphas = eb.InvGamma(df/2, df/2, size=components)
-            pl = eb.t_process(log10_A=log10_A, gamma=gamma, alphas=alphas)
-        elif psd == 'tprocess_adapt':
-            df = 2
-            alpha_adapt = eb.InvGamma(df/2, df/2, size=1)
-            nfreq = parameter.Uniform(-0.5, 10-0.5)
-            pl = eb.t_process_adapt(log10_A=log10_A, gamma=gamma,
-                                    alphas_adapt=alpha_adapt, nfreq=nfreq)
-
-    if psd == 'spectrum':
-        if prior == 'uniform':
-            log10_rho = parameter.LinearExp(-10, -4, size=components)
-        elif prior == 'log-uniform':
-            log10_rho = parameter.Uniform(-10, -4, size=components)
-
-        pl = eb.free_spectrum(log10_rho=log10_rho)
-
-    if select == 'backend':
-        # define selection by observing backend
-        selection = selections.Selection(selections.by_backend)
-    elif select == 'band' or select == 'band+':
-        # define selection by observing band
-        selection = selections.Selection(selections.by_band)
-    else:
-        # define no selection
-        selection = selections.Selection(selections.no_selection)
-
-    if break_flat:
-        log10_A_flat = parameter.Uniform(-20, -11)
-        gamma_flat = parameter.Constant(0)
-        pl_flat = utils.powerlaw(log10_A=log10_A_flat, gamma=gamma_flat)
-
-        freqs = 1.0 * np.arange(1, components+1) / Tspan
-        components_low = sum(f < break_flat_fq for f in freqs)
-        if components_low < 1.5:
-            components_low = 2
-
-        rn = gp_signals.FourierBasisGP(pl, components=components_low, Tspan=Tspan,
-                                       coefficients=coefficients, selection=selection)
-
-        rn_flat = gp_signals.FourierBasisGP(pl_flat, modes=freqs[components_low:],
-                                            coefficients=coefficients, selection=selection,
-                                            name='red_noise_hf')
-        rn = rn + rn_flat
-    else:
-        rn = gp_signals.FourierBasisGP(pl, components=components, Tspan=Tspan,
-                                       coefficients=coefficients, selection=selection,
-                                       modes=modes)
-
-    if select == 'band+': # Add the common component as well
-        rn = rn + gp_signals.FourierBasisGP(pl, components=components, Tspan=Tspan,
-                                   coefficients=coefficients)
-
-    return rn
-
-
-def dm_noise_block(gp_kernel='diag', psd='powerlaw', nondiag_kernel='periodic',
-                   prior='log-uniform', Tspan=None, components=30, gamma_val=None,
-                   coefficients=False):
-    """
-    Returns DM noise model:
-
-        1. DM noise modeled as a power-law with 30 sampling frequencies
-
-    :param psd:
-        PSD function [e.g. powerlaw (default), spectrum, tprocess]
-    :param prior:
-        Prior on log10_A. Default if "log-uniform". Use "uniform" for
-        upper limits.
-    :param Tspan:
-        Sets frequency sampling f_i = i / Tspan. Default will
-        use overall time span for indivicual pulsar.
-    :param components:
-        Number of frequencies in sampling of DM-variations.
-    :param gamma_val:
-        If given, this is the fixed slope of the power-law for
-        powerlaw, turnover, or tprocess DM-variations
-    """
-    # dm noise parameters that are common
-    if gp_kernel == 'diag':
-        if psd in ['powerlaw', 'turnover', 'tprocess', 'tprocess_adapt']:
-            # parameters shared by PSD functions
-            if prior == 'uniform':
-                log10_A_dm = parameter.LinearExp(-20, -11)
-            elif prior == 'log-uniform' and gamma_val is not None:
-                if np.abs(gamma_val - 4.33) < 0.1:
-                    log10_A_dm = parameter.Uniform(-20, -11)
-                else:
-                    log10_A_dm = parameter.Uniform(-20, -11)
-            else:
-                log10_A_dm = parameter.Uniform(-20, -11)
-
-            if gamma_val is not None:
-                gamma_dm = parameter.Constant(gamma_val)
-            else:
-                gamma_dm = parameter.Uniform(0, 7)
-
-            # different PSD function parameters
-            if psd == 'powerlaw':
-                dm_prior = utils.powerlaw(log10_A=log10_A_dm, gamma=gamma_dm)
-            elif psd == 'turnover':
-                kappa_dm = parameter.Uniform(0, 7)
-                lf0_dm = parameter.Uniform(-9, -7)
-                dm_prior = utils.turnover(log10_A=log10_A_dm, gamma=gamma_dm,
-                                          lf0=lf0_dm, kappa=kappa_dm)
-            elif psd == 'tprocess':
-                df = 2
-                alphas_dm = eb.InvGamma(df/2, df/2, size=components)
-                dm_prior = eb.t_process(log10_A=log10_A_dm, gamma=gamma_dm,
-                                        alphas=alphas_dm)
-            elif psd == 'tprocess_adapt':
-                df = 2
-                alpha_adapt_dm = eb.InvGamma(df/2, df/2, size=1)
-                nfreq_dm = parameter.Uniform(-0.5, 10-0.5)
-                dm_prior = eb.t_process_adapt(log10_A=log10_A_dm,
-                                              gamma=gamma_dm,
-                                              alphas_adapt=alpha_adapt_dm,
-                                              nfreq=nfreq_dm)
-
-        if psd == 'spectrum':
-            if prior == 'uniform':
-                log10_rho_dm = parameter.LinearExp(-10, -4, size=components)
-            elif prior == 'log-uniform':
-                log10_rho_dm = parameter.Uniform(-10, -4, size=components)
-
-            dm_prior = eb.free_spectrum(log10_rho=log10_rho_dm)
-
-        dm_basis = utils.createfourierdesignmatrix_dm(nmodes=components,
-                                                      Tspan=Tspan)
-
-    elif gp_kernel == 'nondiag':
-        if nondiag_kernel == 'periodic':
-            # Periodic GP kernel for DM
-            log10_sigma = parameter.Uniform(-10, -4)
-            log10_ell = parameter.Uniform(1, 4)
-            log10_p = parameter.Uniform(-4, 1)
-            log10_gam_p = parameter.Uniform(-3, 2)
-
-            dm_basis = linear_interp_basis_dm(dt=15*86400)
-            dm_prior = periodic_kernel(log10_sigma=log10_sigma, log10_ell=log10_ell,
-                                       log10_gam_p=log10_gam_p, log10_p=log10_p)
-        elif nondiag_kernel == 'periodic_rfband':
-            # Periodic GP kernel for DM with RQ radio-frequency dependence
-            log10_sigma = parameter.Uniform(-10, -4)
-            log10_ell = parameter.Uniform(1, 4)
-            log10_ell2 = parameter.Uniform(2, 7)
-            log10_alpha_wgt = parameter.Uniform(-4, 1)
-            log10_p = parameter.Uniform(-4, 1)
-            log10_gam_p = parameter.Uniform(-3, 2)
-
-            dm_basis = get_tf_quantization_matrix(df=200, dt=15*86400, dm=True)
-            dm_prior = tf_kernel(log10_sigma=log10_sigma, log10_ell=log10_ell,
-                                 log10_gam_p=log10_gam_p, log10_p=log10_p,
-                                 log10_alpha_wgt=log10_alpha_wgt, log10_ell2=log10_ell2)
-        elif nondiag_kernel == 'sq_exp':
-            # squared-exponential GP kernel for DM
-            log10_sigma = parameter.Uniform(-10, -4)
-            log10_ell = parameter.Uniform(1, 4)
-
-            dm_basis = linear_interp_basis_dm(dt=15*86400)
-            dm_prior = se_dm_kernel(log10_sigma=log10_sigma, log10_ell=log10_ell)
-        elif nondiag_kernel == 'sq_exp_rfband':
-            # Sq-Exp GP kernel for DM with RQ radio-frequency dependence
-            log10_sigma = parameter.Uniform(-10, -4)
-            log10_ell = parameter.Uniform(1, 4)
-            log10_ell2 = parameter.Uniform(2, 7)
-            log10_alpha_wgt = parameter.Uniform(-4, 1)
-
-            dm_basis = get_tf_quantization_matrix(df=200, dt=15*86400, dm=True)
-            dm_prior = sf_kernel(log10_sigma=log10_sigma, log10_ell=log10_ell,
-                                 log10_alpha_wgt=log10_alpha_wgt, log10_ell2=log10_ell2)
-        elif nondiag_kernel == 'dmx_like':
-            # DMX-like signal
-            log10_sigma = parameter.Uniform(-10, -4)
-
-            dm_basis = linear_interp_basis_dm(dt=30*86400)
-            dm_prior = dmx_ridge_prior(log10_sigma=log10_sigma)
-
-    dmgp = gp_signals.BasisGP(dm_prior, dm_basis, name='dm_gp',
-                              coefficients=coefficients)
-
-    return dmgp
-
-def scattering_noise_block(kernel='periodic', coefficients=False):
-    """
-    Returns Scattering noise model:
-
-        1. Scattering noise modeled as a power-law with 30 sampling frequencies
-
-    :param psd:
-        PSD function [e.g. powerlaw (default), spectrum, tprocess]
-    :param prior:
-        Prior on log10_A. Default if "log-uniform". Use "uniform" for
-        upper limits.
-    :param Tspan:
-        Sets frequency sampling f_i = i / Tspan. Default will
-        use overall time span for indivicual pulsar.
-    :param components:
-        Number of frequencies in sampling of DM-variations.
-    :param gamma_val:
-        If given, this is the fixed slope of the power-law for
-        powerlaw, turnover, or tprocess DM-variations
-    """
-    #gp_kernel == 'nondiag':
-    if kernel == 'periodic':
-        # Periodic GP kernel for DM
-        log10_sigma = parameter.Uniform(-10, -4)
-        log10_ell = parameter.Uniform(1, 4)
-        log10_p = parameter.Uniform(-4, 1)
-        log10_gam_p = parameter.Uniform(-3, 2)
-
-        dm_basis = linear_interp_basis_scattering(dt=15*86400)
-        dm_prior = periodic_kernel(log10_sigma=log10_sigma, log10_ell=log10_ell,
-                                   log10_gam_p=log10_gam_p, log10_p=log10_p)
-    elif kernel == 'periodic_rfband':
-        # Periodic GP kernel for DM with RQ radio-frequency dependence
-        log10_sigma = parameter.Uniform(-10, -4)
-        log10_ell = parameter.Uniform(1, 4)
-        log10_ell2 = parameter.Uniform(2, 7)
-        log10_alpha_wgt = parameter.Uniform(-4, 1)
-        log10_p = parameter.Uniform(-4, 1)
-        log10_gam_p = parameter.Uniform(-3, 2)
-
-        dm_basis = get_tf_quantization_matrix(df=200, dt=15*86400, dm=True, idx=4)
-        dm_prior = tf_kernel(log10_sigma=log10_sigma, log10_ell=log10_ell,
-                             log10_gam_p=log10_gam_p, log10_p=log10_p,
-                             log10_alpha_wgt=log10_alpha_wgt, log10_ell2=log10_ell2)
-    elif kernel == 'sq_exp':
-        # squared-exponential kernel for DM
-        log10_sigma = parameter.Uniform(-10, -4)
-        log10_ell = parameter.Uniform(1, 4)
-
-        dm_basis = linear_interp_basis_scattering(dt=15*86400)
-        dm_prior = se_dm_kernel(log10_sigma=log10_sigma, log10_ell=log10_ell)
-
-    dmgp = gp_signals.BasisGP(dm_prior, dm_basis, name='scattering_gp',
-                              coefficients=coefficients)
-
-    return dmgp
-
-def dm_annual_signal(idx=2, name='dm_s1yr'):
-    """
-    Returns chromatic annual signal (i.e. TOA advance):
-
-    :param idx:
-        index of radio frequency dependence (i.e. DM is 2). If this is set
-        to 'vary' then the index will vary from 1 - 6
-    :param name: Name of signal
-
-    :return dm1yr:
-        chromatic annual waveform.
-    """
-    log10_Amp_dm1yr = parameter.Uniform(-10, -2)
-    phase_dm1yr = parameter.Uniform(0, 2*np.pi)
-
-    wf = chrom_yearly_sinusoid(log10_Amp=log10_Amp_dm1yr,
-                               phase=phase_dm1yr, idx=idx)
-    dm1yr = deterministic_signals.Deterministic(wf, name=name)
-
-    return dm1yr
-
-def dm_exponential_dip(tmin, tmax, idx=2, sign='negative', name='dmexp'):
-    """
-    Returns chromatic exponential dip (i.e. TOA advance):
-
-    :param tmin, tmax:
-        search window for exponential dip time.
-    :param idx:
-        index of radio frequency dependence (i.e. DM is 2). If this is set
-        to 'vary' then the index will vary from 1 - 6
-    :param sign:
-        set sign of dip: 'positive', 'negative', or 'vary'
-    :param name: Name of signal
-
-    :return dmexp:
-        chromatic exponential dip waveform.
-    """
-    t0_dmexp = parameter.Uniform(tmin,tmax)
-    log10_Amp_dmexp = parameter.Uniform(-10, -2)
-    log10_tau_dmexp = parameter.Uniform(0, 2.5)
-    if sign == 'vary':
-        sign_param = parameter.Uniform(-1.0, 1.0)
-    elif sign == 'positive':
-        sign_param = 1.0
-    else:
-        sign_param = -1.0
-    wf = chrom_exp_decay(log10_Amp=log10_Amp_dmexp,
-                         t0=t0_dmexp, log10_tau=log10_tau_dmexp,
-                         sign_param=sign_param, idx=idx)
-    dmexp = deterministic_signals.Deterministic(wf, name=name)
-
-    return dmexp
-
-def dm_exponential_cusp(tmin, tmax, idx=2, sign='negative', symmetric=False, name='dm_cusp'):
-    """
-    Returns chromatic exponential cusp (i.e. TOA advance):
-
-    :param tmin, tmax:
-        search window for exponential cusp time.
-    :param idx:
-        index of radio frequency dependence (i.e. DM is 2). If this is set
-        to 'vary' then the index will vary from 1 - 6
-    :param sign:
-        set sign of dip: 'positive', 'negative', or 'vary'
-    :param name: Name of signal
-
-    :return dmexp:
-        chromatic exponential dip waveform.
-    """
-    t0_dm_cusp = parameter.Uniform(tmin,tmax)
-    log10_Amp_dm_cusp = parameter.Uniform(-10, -2)
-    log10_tau_dm_cusp_pre = parameter.Uniform(0, 2.5)
-
-    if sign == 'vary':
-        sign_param = parameter.Uniform(-1.0, 1.0)
-    elif sign == 'positive':
-        sign_param = 1.0
-    else:
-        sign_param = -1.0
-
-    if symmetric:
-        log10_tau_dm_cusp_post = 1
-    else:
-        log10_tau_dm_cusp_post = parameter.Uniform(0, 2.5)
-
-    wf = chrom_exp_cusp(log10_Amp=log10_Amp_dm_cusp, sign_param=sign_param,
-                        t0=t0_dm_cusp, log10_tau_pre=log10_tau_dm_cusp_pre,
-                        log10_tau_post=log10_tau_dm_cusp_post, symmetric=symmetric,
-                        idx=idx)
-    dm_cusp = deterministic_signals.Deterministic(wf, name=name)
-
-    return dm_cusp
-
-def dm_dual_exp_cusp(tmin, tmax, idx1=2, idx2=4, sign='negative', symmetric=False, name='dual_dm_cusp'):
-    """
-    Returns chromatic exponential cusp (i.e. TOA advance):
-
-    :param tmin, tmax:
-        search window for exponential cusp time.
-    :param idx:
-        index of radio frequency dependence (i.e. DM is 2). If this is set
-        to 'vary' then the index will vary from 1 - 6
-    :param sign:
-        set sign of dip: 'positive', 'negative', or 'vary'
-    :param name: Name of signal
-
-    :return dmexp:
-        chromatic exponential dip waveform.
-    """
-    t0_dual_cusp = parameter.Uniform(tmin,tmax)
-    log10_Amp_dual_cusp_1 = parameter.Uniform(-10, -2)
-    log10_Amp_dual_cusp_2 = parameter.Uniform(-10, -2)
-    log10_tau_dual_cusp_pre_1 = parameter.Uniform(0, 2.5)
-    log10_tau_dual_cusp_pre_2 = parameter.Uniform(0, 2.5)
-
-    if sign == 'vary':
-        sign_param = parameter.Uniform(-1.0, 1.0)
-    elif sign == 'positive':
-        sign_param = 1.0
-    else:
-        sign_param = -1.0
-
-    if symmetric:
-        log10_tau_dual_cusp_post_1 = 1
-        log10_tau_dual_cusp_post_2 = 1
-    else:
-        log10_tau_dual_cusp_post_1 = parameter.Uniform(0, 2.5)
-        log10_tau_dual_cusp_post_2 = parameter.Uniform(0, 2.5)
-
-    wf = chrom_dual_exp_cusp(t0=t0_dual_cusp, sign_param=sign_param, symmetric=symmetric,
-                        log10_Amp_1=log10_Amp_dual_cusp_1, log10_tau_pre_1=log10_tau_dual_cusp_pre_1,
-                        log10_tau_post_1=log10_tau_dual_cusp_post_1, log10_Amp_2=log10_Amp_dual_cusp_2,
-                        log10_tau_pre_2=log10_tau_dual_cusp_pre_2, log10_tau_post_2=log10_tau_dual_cusp_post_2,
-                        idx1=idx1, idx2=idx2)
-    dm_cusp = deterministic_signals.Deterministic(wf, name=name)
-
-    return dm_cusp
-
-def dmx_signal(dmx_data, name='dmx_signal'):
-    """
-    Returns DMX signal:
-
-    :param dmx_data: dictionary of DMX data for each pulsar from parfile.
-    :param name: Name of signal.
-
-    :return dmx_sig:
-        dmx signal waveform.
-    """
-    dmx = {}
-    for dmx_id in sorted(dmx_data):
-        dmx_data_tmp = dmx_data[dmx_id]
-        dmx.update({dmx_id : parameter.Normal(mu=dmx_data_tmp['DMX_VAL'],
-                                              sigma=dmx_data_tmp['DMX_ERR'])})
-    wf = dmx_delay(dmx_ids=dmx_data, **dmx)
-    dmx_sig = deterministic_signals.Deterministic(wf, name=name)
-
-    return dmx_sig
-
-def chromatic_noise_block(psd='powerlaw', prior='log-uniform', idx=4,
-                          Tspan=None, name='chromatic', components=30,
-                          coefficients=False):
-    """
-    Returns GP chromatic noise model :
-
-        1. Chromatic modeled with user defined PSD with
-        30 sampling frequencies. Available PSDs are
-        ['powerlaw', 'turnover' 'spectrum']
-
-    :param psd:
-        PSD to use for common red noise signal. Available options
-        are ['powerlaw', 'turnover' 'spectrum']
-    :param idx:
-        Index of radio frequency dependence (i.e. DM is 2). If this is set
-        to 'vary' then the index will vary from 1 - 6
-    :param name: Name of signal
-
-    """
-    if psd in ['powerlaw', 'turnover']:
-        if prior == 'uniform':
-            log10_A = parameter.LinearExp(-18, -11)
-        elif prior == 'log-uniform':
-            log10_A = parameter.Uniform(-18, -11)
-        gamma = parameter.Uniform(0, 7)
-
-        # PSD
-        if psd == 'powerlaw':
-            cpl = utils.powerlaw(log10_A=log10_A, gamma=gamma)
-        elif psd == 'turnover':
-            kappa = parameter.Uniform(0, 7)
-            lf0 = parameter.Uniform(-9, -7)
-            cpl = utils.turnover(log10_A=log10_A, gamma=gamma,
-                                 lf0=lf0, kappa=kappa)
-
-    if psd == 'spectrum':
-        if prior == 'uniform':
-            log10_rho = parameter.LinearExp(-10, -4, size=components)
-        elif prior == 'log-uniform':
-            log10_rho = parameter.Uniform(-10, -4, size=components)
-        cpl = eb.free_spectrum(log10_rho=log10_rho)
-
-    # set up signal
-    # JS: This does not work with basis_quad function below
-    #if idx == 'vary':
-    #    c_idx = parameter.Uniform(0, 6)
-
-    # quadratic piece
-    basis_quad = chromatic_quad_basis(idx=idx)
-    prior_quad = chromatic_quad_prior()
-    cquad = gp_signals.BasisGP(prior_quad, basis_quad, name=name+'_quad')
-
-    # Fourier piece
-    basis_gp = eb.createfourierdesignmatrix_chromatic(nmodes=components,
-                                                     Tspan=Tspan)
-    cgp = gp_signals.BasisGP(cpl, basis_gp, name=name+'_gp',
-                             coefficients=coefficients)
-
-    return cquad + cgp
-
-def common_red_noise_block(psd='powerlaw', prior='log-uniform',
-                           Tspan=None, components=30, gamma_val=None,
-                           orf=None, name='gw', coefficients=False):
-    """
-    Returns common red noise model:
-
-        1. Red noise modeled with user defined PSD with
-        30 sampling frequencies. Available PSDs are
-        ['powerlaw', 'turnover' 'spectrum']
-
-    :param psd:
-        PSD to use for common red noise signal. Available options
-        are ['powerlaw', 'turnover' 'spectrum']
-    :param prior:
-        Prior on log10_A. Default if "log-uniform". Use "uniform" for
-        upper limits.
-    :param Tspan:
-        Sets frequency sampling f_i = i / Tspan. Default will
-        use overall time span for indivicual pulsar.
-    :param gamma_val:
-        Value of spectral index for power-law and turnover
-        models. By default spectral index is varied of range [0,7]
-    :param orf:
-        String representing which overlap reduction function to use.
-        By default we do not use any spatial correlations. Permitted
-        values are ['hd', 'dipole', 'monopole'].
-    :param name: Name of common red process
-
-    """
-
-    orfs = {'hd': utils.hd_orf(), 'dipole': utils.dipole_orf(),
-            'monopole': utils.monopole_orf()}
-
-    # common red noise parameters
-    if psd in ['powerlaw', 'turnover', 'turnover_knee']:
-        amp_name = '{}_log10_A'.format(name)
-        if prior == 'uniform':
-            log10_Agw = parameter.LinearExp(-18, -11)(amp_name)
-        elif prior == 'log-uniform' and gamma_val is not None:
-            if np.abs(gamma_val - 4.33) < 0.1:
-                log10_Agw = parameter.Uniform(-18, -14)(amp_name)
-            else:
-                log10_Agw = parameter.Uniform(-18, -11)(amp_name)
-        else:
-            log10_Agw = parameter.Uniform(-18, -11)(amp_name)
-
-        gam_name = '{}_gamma'.format(name)
-        if gamma_val is not None:
-            gamma_gw = parameter.Constant(gamma_val)(gam_name)
-        else:
-            gamma_gw = parameter.Uniform(0, 7)(gam_name)
-
-        # common red noise PSD
-        if psd == 'powerlaw':
-            cpl = utils.powerlaw(log10_A=log10_Agw, gamma=gamma_gw)
-        elif psd == 'turnover':
-            kappa_name = '{}_kappa'.format(name)
-            lf0_name = '{}_log10_fbend'.format(name)
-            kappa_gw = parameter.Uniform(0, 7)(kappa_name)
-            lf0_gw = parameter.Uniform(-9, -7)(lf0_name)
-            cpl = utils.turnover(log10_A=log10_Agw, gamma=gamma_gw,
-                                 lf0=lf0_gw, kappa=kappa_gw)
-        elif psd == 'turnover_knee':
-            kappa_name = '{}_kappa'.format(name)
-            lfb_name = '{}_log10_fbend'.format(name)
-            delta_name = '{}_delta'.format(name)
-            lfk_name = '{}_log10_fknee'.format(name)
-            kappa_gw = parameter.Uniform(0, 7)(kappa_name)
-            lfb_gw = parameter.Uniform(-9.3, -8)(lfb_name)
-            delta_gw = parameter.Uniform(-2, 0)(delta_name)
-            lfk_gw = parameter.Uniform(-8, -7)(lfk_name)
-            cpl = eb.turnover_knee(log10_A=log10_Agw, gamma=gamma_gw,
-                                   lfb=lfb_gw, lfk=lfk_gw,
-                                   kappa=kappa_gw, delta=delta_gw)
-
-    if psd == 'spectrum':
-        rho_name = '{}_log10_rho'.format(name)
-        if prior == 'uniform':
-            log10_rho_gw = parameter.LinearExp(-9, -4, size=components)(rho_name)
-        elif prior == 'log-uniform':
-            log10_rho_gw = parameter.Uniform(-9, -4, size=components)(rho_name)
-
-        cpl = eb.free_spectrum(log10_rho=log10_rho_gw)
-
-    if orf is None:
-        crn = gp_signals.FourierBasisGP(cpl, coefficients=coefficients,
-                                        components=components, Tspan=Tspan,
-                                        name=name)
-    elif orf in orfs.keys():
-        crn = gp_signals.FourierBasisCommonGP(cpl, orfs[orf], #coefficients=coefficients,
-                                              components=components, Tspan=Tspan,
-                                              name=name)
-    else:
-        raise ValueError('ORF {} not recognized'.format(orf))
-
-    return crn
-
 
 ###############################
 ###  PTA models from paper  ###
